fixes from CR
